# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AEO Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prompts/simulate-all:
    post:
      operationId: simulateAllPrompts
      tags: [tracking]
      summary: Run every active prompt for a company across its configured models
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateAllBody"
      responses:
        "200":
          description: Batch result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingJob"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "409":
          description: A batch is already running for this company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "503":
          description: Model API key not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /prompts/{id}/simulate-all:
    post:
      operationId: simulatePromptAllModels
      tags: [tracking]
      summary: Run one prompt across every enabled registry model as a durable server-side job
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Job started (poll /tracking/jobs?promptId=)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingJob"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "404":
          description: Prompt not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "409":
          description: A run is already in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "503":
          description: Model API key not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /tracking/jobs/{id}/attempts:
    get:
      operationId: getTrackingJobAttempts
      tags: [tracking]
      summary: Per-model attempt outcomes for a job (exact partial-failure reporting)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Attempt counts grouped by model
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobModelAttempts"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
  /tracking/jobs:
    get:
      operationId: listTrackingJobs
      tags: [tracking]
      summary: Recent bulk tracking runs for a company
      parameters:
        - name: companyId
          in: query
          required: true
          schema: { type: integer }
        - name: promptId
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Jobs, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackingJob"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
components:
...
  schemas:
...
    JobModelSnapshot:
      type: object
      required: [id, label, provider, supportsSearch]
      properties:
        id: { type: string }
        label: { type: string }
        provider: { type: string }
        supportsSearch: { type: boolean }
    JobModelAttempts:
      type: object
      required: [model, total, succeeded, failed]
      properties:
        model: { type: string }
        total: { type: integer }
        succeeded: { type: integer }
        failed: { type: integer }
        lastError: { type: ["string", "null"] }
    ErrorMessage:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    SimulateAllBody:
      type: object
      properties:
        companyId: { type: integer }
      required: [companyId]
    TrackingJob:
      type: object
      properties:
        id: { type: integer }
        companyId: { type: integer }
        promptId: { type: ["integer", "null"] }
        models:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/JobModelSnapshot"
        runDate: { type: string }
        trigger: { type: string, enum: [manual, scheduled] }
        status: { type: string, enum: [running, completed, failed] }
        totalSimulations: { type: integer }
        succeeded: { type: integer }
        failed: { type: integer }
        error: { type: ["string", "null"] }
        createdAt: { type: string }
        finishedAt: { type: ["string", "null"] }
      required:
        [id, companyId, runDate, trigger, status, totalSimulations, succeeded, failed, createdAt]
...